test_Booking/requirement.txt
pytest
pytest-xdist
requests
allure-pytest
behave
allure-behave